+++ packages/backtest/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @zapcli/backtest
 
+## 1.0.12
+
+### Patch Changes
+
+- change reports for dd and positions
+- Updated dependencies
+  - @zapcli/core@1.0.12
+
 ## 1.0.11
 
 ### Patch Changes